requirements.txt
git+https://github.com/tadashi-aikawa/chalice.git
requests==2.12.1
boto3==1.4.1
pydash==3.4.7
pytz==2016.7
typing==3.5.2.2
git+https://github.com/tadashi-aikawa/dictmixin
pytest==3.0.4
pytest-cov==2.4.0
pyyaml==3.12
